Uplift some feeding out of associated_type_for_impl_trait_in_impl and into queries
